Fixes current limit on PositionJoints
-makes PositionJointConfig.currentlimit an integer
-applies that in IOSparkMax
- TalonFX will not need changes due to implicit conversion

+++ src/main/java/frc/robot/subsystems/position_joint/PositionJointConstants.java
@@ -68,7 +68,7 @@ public class PositionJointConstants {
       boolean[] reversed,
       double gearRatio,
       double momentOfInertiaKgMetersSquared,
-      double currentLimit,
+      int currentLimit,
       GravityType gravity,
       EncoderType encoderType,
       int encoderID,
